fixed all warning

+++ Project/DataModels/RoleLevelModel.cs
@@ -1,14 +1,14 @@
 public class RoleLevelModel
 {
-    public Int64 Id { get; set; }
+    public int Id { get; set; }
     public string Functionalty { get; set; }
-    public Int64 LevelNeeded { get; set; }
+    public int LevelNeeded { get; set; }
 
     public RoleLevelModel(Int64 id, string functionality, Int64 level_Needed)
     {
-        Id = id;
+        Id = (int)id;
         Functionalty = functionality;
-        LevelNeeded = level_Needed;
+        LevelNeeded = (int)level_Needed;
     }
 
     public RoleLevelModel(string functionalty, int level_Needed)
